Update client.lua
fixed an issue that was causing peds to keep proofs when not wearing the respective gear.  fixed an issue that was causing peds to continue to be able to turn gear lights on and off if they werent wearing gear equipped with a light.  fixed an issue that caused gear lights to stay on and illuminated after changing out of gear.